fix mirador app not being unmounted when navigating
This should fix #14.

Mirador generates class names for its components. They are named
mirador1, ..., miradorN.

These classes have associated css rules:
```
.mirador50 {
  display: flex;
}
```

If we unmount our mirador vue component without unmounting the Mirador
app, these css rules are not removed from the DOM.

The render order of the Mirador app's component impacts the naming of
the generated classes, so we can end up in a situation where an element is associated
to both the correct rule and the rule associated with the previous
element which had this class name, leading to bad rendering.